Inline all code into .hh file and eliminate std calls
