+++ src/orange/CMakeLists.txt
@@ -23,7 +23,6 @@ list(APPEND SOURCES
   detail/BIHPartitioner.cc
   detail/DepthCalculator.cc
   detail/OrangeInputIOImpl.json.cc
-  detail/OrientedBoundingZone.cc
   detail/RectArrayInserter.cc
   detail/SurfacesRecordBuilder.cc
   detail/UnitInserter.cc
